+++ src/plant/plant.service.ts
@@ -4,6 +4,7 @@ import {
   HttpStatus,
   Injectable,
 } from '@nestjs/common';
+import { CloudinaryService } from 'src/cloudinary/cloudinary.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { RedisCacheService } from 'src/redis-cache/redis-cache.service';
 import { PlantDto, UpdatePlantDto } from './dto';
@@ -12,10 +13,13 @@ import { PlantType } from './types';
 @Injectable()
 export class PlantService {
   private readonly prefix: string = 'plants:';
+  private readonly folder: string = 'plants';
+  private readonly urlPrefix: string = 'plants/';
 
   constructor(
     private prisma: PrismaService,
     private cache: RedisCacheService,
+    private cloudinary: CloudinaryService,
   ) {}
 
   async getPlants(url: string): Promise<PlantType[]> {
@@ -91,7 +95,19 @@ export class PlantService {
 
   async createPlant(userId: string, dto: PlantDto): Promise<PlantType> {
     try {
-      const plant: PlantType = await this.prisma.plant.create({
+      const plant: PlantType = {
+        ...dto,
+        photo: await this.cloudinary.upload(dto.photo, this.folder),
+        user_id: userId,
+      };
+
+      if (!plant.photo)
+        throw new HttpException(
+          'Could not decode base64',
+          HttpStatus.EXPECTATION_FAILED,
+        );
+
+      const storedPlant: PlantType = await this.prisma.plant.create({
         data: {
           ...dto,
           user_id: userId,
@@ -100,7 +116,7 @@ export class PlantService {
 
       await this.cache.del(this.prefix);
 
-      return plant;
+      return storedPlant;
     } catch (error) {
       throw error;
     }
@@ -112,18 +128,39 @@ export class PlantService {
     dto: UpdatePlantDto,
   ): Promise<PlantType> {
     try {
-      const plant = await this.prisma.plant.findUnique({
+      const storedPlant = await this.prisma.plant.findUnique({
         where: { id },
       });
 
-      if (!plant || plant.user_id !== userId)
+      if (!storedPlant || storedPlant.user_id !== userId)
         throw new ForbiddenException('Access to ressources denied');
+
+      const cloudDelete = await this.cloudinary.delete(
+        storedPlant.photo,
+        this.urlPrefix,
+      );
+
+      if (cloudDelete.result !== 'ok')
+        throw new HttpException(
+          'error delete cloudinary img !',
+          HttpStatus.EXPECTATION_FAILED,
+        );
+
+      const photo: string = await this.cloudinary.upload(
+        dto.photo,
+        this.folder,
+      );
+
+      const data: UpdatePlantDto = {
+        ...dto,
+        photo,
+      };
 
       await this.cache.del(this.prefix);
 
       return this.prisma.plant.update({
         where: { id },
-        data: { ...dto },
+        data: { ...data },
       });
     } catch (error) {
       throw error;
@@ -138,6 +175,17 @@ export class PlantService {
 
       if (!plant || plant.user_id !== userId)
         throw new ForbiddenException('Access to ressources denied');
+
+      const cloudDelete = await this.cloudinary.delete(
+        plant.photo,
+        this.urlPrefix,
+      );
+
+      if (cloudDelete.result !== 'ok')
+        throw new HttpException(
+          'error delete cloudinary img !',
+          HttpStatus.EXPECTATION_FAILED,
+        );
 
       await this.prisma.plant.delete({
         where: { id },
